ITPW/requirements.txt
Django==4.2.19
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.2.0  # For image handling
